requirements.txt
python-dotenv~=1.0.1
requests~=2.31.0
pip~=24.0
aiogram~=3.6.0
